requirements.txt
Flask==2.1.3
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
gunicorn
marshmallow==3.17.0
psycopg2-binary
PyJWT==2.4.0
python-dotenv==0.20.0
requests==2.28.1